Database integrated + intrusions
