--- requirements.txt
@@ -1,7 +1,0 @@
-fastapi
-uvicorn
-starlette
-pydantic
-sqlalchemy
-psycopg2-binary
-postgresql
